+++ src/main/java/kitchenpos/application/TableGroupService.java
@@ -3,8 +3,8 @@ package kitchenpos.application;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.NoSuchElementException;
-import kitchenpos.domain.Order;
-import kitchenpos.domain.OrderRepository;
+import kitchenpos.order.domain.Order;
+import kitchenpos.order.domain.OrderRepository;
 import kitchenpos.domain.OrderTable;
 import kitchenpos.domain.OrderTableRepository;
 import kitchenpos.domain.TableGroup;
