+++ src/main/java/kitchenpos/order/dto/OrderLineItemRequest.java
@@ -1,6 +1,6 @@
-package kitchenpos.dto;
+package kitchenpos.order.dto;
 
-import kitchenpos.domain.OrderLineItem;
+import kitchenpos.order.domain.OrderLineItem;
 
 public class OrderLineItemRequest {
