+++ src/main/java/kitchenpos/order/dto/OrderResponse.java
@@ -1,10 +1,10 @@
-package kitchenpos.dto;
+package kitchenpos.order.dto;
 
 import java.time.LocalDateTime;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.stream.Collectors;
-import kitchenpos.domain.Order;
+import kitchenpos.order.domain.Order;
 
 public class OrderResponse {
