+++ src/test/java/kitchenpos/domain/OrderLineItemTest.java
@@ -3,6 +3,8 @@ package kitchenpos.domain;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
+import kitchenpos.order.domain.Order;
+import kitchenpos.order.domain.OrderLineItem;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
